fix: proper evaluation if cta is set
the outer with runs false if the whole section call_to_action is missing in the config, the inner if then checks of the section has enable set. This way we can write in the documentation that to remove that section permanently the user can just remove the whole section block instead of just setting .enable to false. The other way around the with would be always true, if call_to_action.enable exists. If that does not exist a notice/error would have been thrown.